more ui.message refinement
